setup.py: Run make before relevant commands
It seems that, unless the package_data files are present before the
commands are invoked, they will not be included in the distribution
files. So, we override the build_py, sdist, and bdist_wheel commands,
and run 'make' beforehand. We might need to add additional commands to
this list in the future.

This fixes #77