Rewrite run_nosetests.sh in Python
This allows me to add a --no-local-import option, which removes the
current directory from sys.path before invoking nose. By doing so, this
allows the CI job to test the installed version of scuba, and not the
local version, essentially verifying that the build/install process
works correctly.

run_full_tests.py already invokes scuba without a path, relying on the
fact that it is installed.