scuba: Treat "scripts" as lists of lines, not lists of lists
We want to leave the scripts largely untouched, to enable proper
operation of shell constructs like this:

 - cd somewhere && ./run.sh

By using shlex_split() when parsing and shell_quote_cmd() when writing
the script, we were causing things like && to become quoted and lose their
magic.

This changes ScubaAlias.from_dict() and ScubaConfig.process_command() to
treat processed scripts as lists of command-line strings, rather than
lists of lists. This requires shell_quote_cmd() to be used in places where
we are adding to the command line strings (e.g. entrypoint, and user args).

Note that this also required updating all process_command() tests; the
results are considered as sh would consider them: by splitting and
verifying the elements are as expected.

Fixes #100