Update scuba to work on Windows Subsystem for Linux
The reason scuba previously didn't work with WSL is two part, WSL must
connect to docker remotely, and the file paths that must be supplied
for WSL to work with docker is different than conventional Linux file
paths, as well as the file paths used by scuba. By having scuba check if
it is running from within WSL it is able to verify that the proper paths
are set up (Typically by mounting C:\ as /c/) and that the files needed
to run (scubainit and the temp directory) are placed in a location that
Windows is able to find (C:\ProgramData\Scuba or /c/ProgramData/Scuba in
WSL). These changes allow scuba to run on WSL which is connecting to
Docker remotely.