scuba: Don't run the entrypoint for each line in a multi-line script
**This is a breaking change.**

The previous logic was generally incorrect. For a multi-line script,
Scuba would prepend each line with the entrypoint. This sort of makes
sense, except the script is run within *one*  `docker run` invocation.
Many image entrypoints are wrappers which set up things in the
container, and then execute the user command which is passed as
arguments to the entrypoint (e.g. `exec "@"`). Thus, it isn't right to
execute the entrypoint multiple times in the same container.

This changes the behavior of Scuba to instead invoke the entrypoint,
passing it the path of the scuba-generated script to be executed. In
this way, the entrypoint is only run once.

N.B. This means that Scuba no longer works with images which wrap a
single command (e.g. an "application" image which sets some CLI
application as the entrypoint). If this use case is needed in the
future, an additional YAML option (e.g. `entrypoint_each_line`) could be
added to restore the old behavior.

Fixes #114