Remove unnecessary shebang from scripts invoked explicitly with a shell
Also, change the way that the unit tests validate which shell is running
in the container to be more rigorous. Previously, a change to the
location of scuba's command.sh file would have broken tests. Also, an
error where the invoking shell was changed but not the shebang may not
have been caught by the tests.