scubainit: Ignore matching passwd/group/shadow entries
Commit 254fc4a7 (#139) relaxed the checks for existing passwd entries
matching the UID being added (and group, GID).

This relaxes those checks even further by allowing existing entries
which match both name *and* ID. We can't ignore the checks altogether
because if we added a duplicate name with a different ID, the name would
become ambiguous.

The only real use case I can think of here is for "root" but I suppose
there could be others, and this seems correct.