Add images completer and update aliases completer
Per discussions on JonathonReinhart/scuba#162

An images completer is now present to list all available Docker images
for the `--image` option.
Fixes up the aliases completer to *not* give a suggestion when an alias
has already been specified and to warn the user if no config file is
present.
Removed `get_paths_and_config()` and updated `find_config()` to return
the loaded config.
Fixes up and adds unit tests.