ci: Don't run workflows on push
Ideally we would run on push, but that doesn't cover when someone
creates/updates a PR on a fork, so we added pull_request in cf00c12f.
That works well for forks.

However, that means when I push to a branch on my project with an
associated PR, the tests all run twice.

Pushing to a branch without a PR, and expecting a build to run, seems
unlikely enough that we can use pull_request exclusively for now.