scubainit: Always chown home directory
Prior to this change, if the home directory already existed (because the
scuba workdir was bind-mounted somewhere underneath it), then the home
directory would remain owned by root.

Removing this check ensures that the chown() always happens.

The only odd thing remaining is that any other directories from $HOME
down to the workdir will remain owned as root.

This fixes #168