Allow docker arguments in scubafile
Docker arguments can now be specified in the scubafile at the top-level
or per-alias. The order of precendence for a single run is:

1. Arguments provided on command line via `-d`
2. Per-alias `docker_args` key
2. Top-level `docker_args` key

Explicitly providing a null or empty string value for `docker_args` at
the alias level overrides the top-level `docker_args`, and results in no
extra args passed to docker.