Fix missing import in tests
While working on this feature, I hit a case where the 'warnings' module
was referenced, but this caused a NameError because the module was never
imported. AFAICT this issue has been present since the code was
introduced, but it was never noticed because it was behind a conditional
that was never satisfied.