Re-write env variable expansion to be more general-purpose and correct
Responding to a comment on GH-192, the code now calls
string.Template.substitute, which has more correct behavior because it
won't skip expanding any unset environmental variables. Now, the
behavior is to raise an exception, similar to bash's nounset (set -u)
option.

This code was placed in 'utils.py' so that it can be used elsewhere in
Scuba and not just config-parsing code.