scuba: Fix make_vol_opt trailing colon bug
When make_vol_opt() was called with options=[], it would return a volume
spec which erroneously contained a trailing colon.

Also, this removes the unnecessary flexibility of passing a single
option string (as opposed to a sequence of option strings). This wasn't
intended by the type signature for options, but unforunately a 'str'
is a 'Sequence[str]'. See https://github.com/python/mypy/issues/11001.

This fixes #215.