scuba: Plumb scuba_root through config functions
Starting from load_config(), this passes scuba_root (the directory
in which .scuba.yml is found) through the hierarchy to
ScubaVolume.from_dict().

This will enable us to use scuba_root in config value expansion,
particularly volume paths.

Note that scuba_root is marked Optional[Path] to support an empty or
default ScubaConfig, or one that is programatically created without a
scuba_root directory. This must be accounted for when it is used.