scubainit: Avoid lifetimes on EntFile{Reader,Writer}
Rather than use lifetimes on the EntFileReader and EntFileWriter
structs, we can implement the common into_inner() convention. This
consumes the outer object returning the inner wrapped object, which in
this case is the File which we want to reuse. This matches e.g.
`std::io::BufReader::into_inner()`.

Note that this also requires changing the way the reader is iterated: We
use `for x in &mut reader` which avoids moving the `reader` object into
the loop, allowing `into_inner()` to be subsequently called. See the
discussion here:
https://github.com/JonathonReinhart/scuba/pull/239#issuecomment-1872970751.