WIP fixing #219
This works if scubaroot = /home/$USER: is_bind_mount() detects the bind mount
because we're trying to mount the /jreinhart subdirectory of the /home/
mount:

  scubainit: ERROR - Error creating home dir: "/home/jreinhart" is on a bind-mount: /dev/mapper/magnum--vg-home[/jreinhart]

But it fails for scubaroot = /home: is_bind_mount() sees root="/" and
returns false.

I think there is simply no way to reliably detect a bind mount!

So now we might be to the point of either passing the list of
bind-mount volumes from scuba to scubainit, or by checking for this
situation in scuba instead.