scuba: Default _get_image_config() to return None if Config key is missing
Docker Engine 28.2 changed the output of 'docker inspect --type image'
to no longer include keys when their values are 'null'. This could cause
get_image_command() and get_image_entrypoint() to raise an exception if
the Cmd or Entrypoint keys are missing, respectively.

Change from [] to .get() to return None if the key is missing, restoring
the previous behavior.

Fixes #262